app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
requests==2.31.0